BAU: dont include PII in error messages
This commit removes a users full email from the error message returned
through the API.

Consumers of this endpoint currently log the response from the HTTP
call which can then propegate through a number of systems.

Include the domain of the accounts email address to keep parity
with being able to identify if the account failing to be modified
looks like a test account/ script or a genuine user.

The preferred method of resolving this would be to specify a shared
constant as an `error_code` (for example) so that the consumer of the
API can check for this failure in particular and log accordingly. As
there doesn't seem to be precedent for that currently just resolve
this at the producer end.